Room Service done

+++ viberbot/src/main/java/com/one/viberbot/service/RoomService.java
@@ -7,7 +7,7 @@ public interface RoomService {
 	public Iterable<Room> findAll();
 	public void add(Room room);
 	public void update(Room room);
-	public Room getOne(Integer id);
-	public void delete(Integer id);
+	public Room getOne(Long id);
+	public void delete(Long id);
 	
 }
